+++ app/Http/Controllers/Admin/AdminQuestionActionController.php
@@ -25,6 +25,7 @@
 
 namespace Gamify\Http\Controllers\Admin;
 
+use Gamify\Enums\QuestionActuators;
 use Gamify\Http\Requests\QuestionActionCreateRequest;
 use Gamify\Models\Question;
 use Gamify\Models\QuestionAction;
@@ -34,9 +35,9 @@ class AdminQuestionActionController extends AdminController
     /**
      * Show the form for creating a new resource.
      *
-     * @param Question $question
+     * @param  Question  $question
      *
-     * @return \Illuminate\View\View
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\View\View
      */
     public function create(Question $question)
     {
@@ -47,20 +48,24 @@ class AdminQuestionActionController extends AdminController
             $availableActions[$action->id] = $action->name;
         }
 
-        return view('admin/action/create', compact('question', 'availableActions'));
+        return view('admin/action/create', [
+            'question' => $question,
+            'availableActions' => $availableActions,
+            'actuators' => QuestionActuators::asSelectArray(),
+        ]);
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param Question                    $question
-     * @param QuestionActionCreateRequest $request
+     * @param  Question  $question
+     * @param  QuestionActionCreateRequest  $request
      *
      * @return \Illuminate\Http\RedirectResponse
      */
     public function store(Question $question, QuestionActionCreateRequest $request)
     {
-        $question->actions()->create($request->all());
+        $question->actions()->create($request->validated());
 
         return redirect()->route('admin.questions.edit', $question)
             ->with('success', __('admin/action/messages.create.success'));
@@ -69,12 +74,12 @@ class AdminQuestionActionController extends AdminController
     /**
      * Remove the specified resource from storage.
      *
-     * @param Question       $question
-     * @param QuestionAction $action
-     *
-     * @throws \Exception
+     * @param  Question  $question
+     * @param  QuestionAction  $action
      *
      * @return \Illuminate\Http\RedirectResponse
+     * @throws \Exception
+     *
      */
     public function destroy(Question $question, QuestionAction $action)
     {